pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "dir-sync"
version = "0.1.0"
description = "Cross-platform toolbar app that synchronizes directories via rsync-style strategies."
authors = [{ name = "Dir Sync Contributors" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "psutil>=5.9",
    "pystray>=0.19",
    "pillow>=10.0",
    "plyer>=2.1",
    "typer>=0.12",
    "croniter>=1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=24.3",
    "ruff>=0.4",
]

[project.scripts]
dir-sync = "dirsync.app:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dirsync"]

[tool.hatch.build.targets.sdist]
include = ["src/dirsync", "pyproject.toml", "README.md", "CHANGELOG.md"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
